+++ modules/npm-release/mock-package.json
@@ -1,0 +1,5 @@
+{
+  "name": "@test-mock/test-mock",
+  "version": "0.0.10",
+  "description": "This is purely a testmock"
+}
