--- modules/server/body-parser-middleware.test.ts
@@ -1,56 +1,0 @@
-import { describe, expect, it, jest, spyOn } from "bun:test";
-import { bodyParser, getParsedBody } from "./body-parser-middleware";
-
-describe("bodyParser middleware", () => {
-  it("should correctly parse JSON content", async () => {
-    const mockRequest = {
-      headers: new Map([["Content-Type", "application/json"]]),
-      text: jest.fn().mockResolvedValue(JSON.stringify({ key: "value" })),
-    } as unknown as Request;
-
-    const mockNext = jest.fn();
-
-    const response = new Response();
-
-    await bodyParser({ request: mockRequest, next: mockNext, response });
-
-    expect(getParsedBody<{ key: string }>(mockRequest)).toEqual({
-      key: "value",
-    });
-    expect(mockNext).toHaveBeenCalled();
-  });
-
-  it("should handle non-JSON content", async () => {
-    const mockRequest = {
-      headers: new Map([["Content-Type", "text/plain"]]),
-      text: jest.fn().mockResolvedValue("plain text"),
-    } as unknown as Request;
-
-    const mockNext = jest.fn();
-    const response = new Response();
-
-    await bodyParser({ request: mockRequest, next: mockNext, response });
-
-    expect(getParsedBody<string>(mockRequest)).toBe("plain text");
-    expect(mockNext).toHaveBeenCalled();
-  });
-
-  it("should handle invalid JSON gracefully", async () => {
-    const consoleErrorSpy = spyOn(console, "error").mockImplementation(
-      () => ({})
-    );
-    const mockRequest = {
-      headers: new Map([["Content-Type", "application/json"]]),
-      text: jest.fn().mockResolvedValue("invalid json"),
-    } as unknown as Request;
-
-    const mockNext = jest.fn();
-    const response = new Response();
-
-    await bodyParser({ request: mockRequest, next: mockNext, response });
-
-    expect(consoleErrorSpy).toHaveBeenCalled();
-    expect(mockNext).toHaveBeenCalled();
-    consoleErrorSpy.mockRestore();
-  });
-});
