+++ modules/test-utils/index.ts
@@ -1,0 +1,1 @@
+export { isTestFile } from "./test-utils";
