+++ modules/test-utils/test-utils.ts
@@ -1,0 +1,11 @@
+export const isTestFile = (
+  meta: ImportMeta,
+  testFileMatch: string = "test"
+) => {
+  console.log({ import: meta });
+
+  const splitFile = meta.file.split(".");
+  const fileExtensions = splitFile.slice(1).join(".");
+
+  return fileExtensions?.includes(testFileMatch) || false;
+};
